Tests/LibWeb/Text/expected/hit_testing/visibility-visible-inside-hidden-stacking-context.txt

<#text>
<DIV id="visible-child">
